+++ synth-sample-and-hold.h
@@ -26,7 +26,7 @@ namespace SFM
 
 		~SampleAndHold() {}
 
-		SFM_INLINE float Sample(double phase, float input)
+		SFM_INLINE float Sample(float phase, float input)
 		{
 			const float curGate = oscSquare(phase);
